+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and what APIs have changed, if applicable.
 
 ## [Unreleased]
 
+## [29.8.3] - 2020-11-09
+- Support symbolTable requests with suffixes
+
 ## [29.8.2] - 2020-11-06
 - Fix bug: if there is no input schema, do not run pegasusSchemaSnapshotCheck. The check statement was wrong.
 
@@ -4736,7 +4739,8 @@ patch operations can re-use these classes for generating patch messages.
 
 ## [0.14.1]
 
-[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.8.2...master
+[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.8.3...master
+[29.8.3]: https://github.com/linkedin/rest.li/compare/v29.8.2...v29.8.3
 [29.8.2]: https://github.com/linkedin/rest.li/compare/v29.8.1...v29.8.2
 [29.8.1]: https://github.com/linkedin/rest.li/compare/v29.8.0...v29.8.1
 [29.8.0]: https://github.com/linkedin/rest.li/compare/v29.7.15...v29.8.0